refactor: ES Module に変更し，ドメインに関係ない関数を別 Module に分離
