+++ index.html
@@ -3,7 +3,7 @@
     <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
     <title>神経衰弱ゲーム</title>
     <link rel="stylesheet" href="main.css">
-    <script type="text/javascript" src="index.js"></script>
+    <script type="module" src="index.mjs"></script>
 </head>
 <body>
     <div id="panel"></div>
